+++ app/Models/Adicional.php
@@ -4,6 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Adicional extends Model
 {
@@ -15,4 +16,9 @@ class Adicional extends Model
         'evento_id',
         'creador_id',
     ];
+
+    public function adicional_cache(): HasMany
+    {
+        return $this->hasMany(Adicional_Cache::class, 'adicional_id');
+    }
 }